pyproject.toml
[project]
name = "data-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "e2b-code-interpreter>=2.4.1",
    "google-genai>=1.55.0",
    "google-generativeai>=0.8.6",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.1",
    "tabulate>=0.9.0",
]
